fix: move calculations to chart component
