+++ src/components/ContributionByProvinceToGoal/ContributionByProvinceToGoal.vue
@@ -6,12 +6,14 @@
 </template>
 
 <script>
+import layers from "@/data/administrative-boundaries-layers";
 import { use } from "echarts/core";
 import { CanvasRenderer } from "echarts/renderers";
 import { PieChart } from "echarts/charts";
 import { TooltipComponent } from "echarts/components";
 import VChart from "vue-echarts";
 import DownloadChartButton from "../DownloadChartButton/DownloadChartButton.vue";
+import buildFeatureUrl from "@/lib/build-feature-url";
 
 use([CanvasRenderer, PieChart, TooltipComponent]);
 
@@ -19,13 +21,6 @@ export default {
   components: {
     VChart,
     DownloadChartButton,
-  },
-
-  props: {
-    data: {
-      type: Array,
-      required: true,
-    },
   },
 
   data() {
@@ -43,18 +38,55 @@ export default {
           },
         ],
       },
+      layer: layers.find((layer) => layer.id === "level-1"),
     };
   },
 
-  watch: {
-    data(value) {
-      this.updateChart(value);
-    },
-  },
-
   methods: {
-    updateChart(data) {
-      this.option.series[0].data = data;
+    async updateChart() {
+      const response = await fetch(
+        buildFeatureUrl({
+          ...this.layer,
+          propertyName: "name_1,mg_area2020",
+        })
+      );
+      const data = await response.json();
+
+      this.option.series[0].data = this.getContributionByProvince(data);
+    },
+    getContributionByProvince(data) {
+      let totalRestoredArea = 0;
+
+      const contributionByProvince = data?.features.map((feature) => {
+        const area = feature?.properties?.mg_area2020;
+        totalRestoredArea += area;
+
+        return {
+          name: feature?.properties?.name_1,
+          value: this.getPercentage(area, 600000),
+        };
+      });
+
+      const toRestoreArea = this.getPercentage(
+        600000 - totalRestoredArea,
+        600000
+      );
+
+      contributionByProvince.push({
+        value: toRestoreArea,
+        name: "To Restore",
+        itemStyle: {
+          color: "#D3D3D3",
+        },
+      });
+
+      return contributionByProvince;
+    },
+    getPercentage(value, total) {
+      return (value / total) * 100;
+    },
+    resetChartData() {
+      this.option.series[0].data = [];
     },
     resizeChart() {
       this.$refs.chart?.resize();
@@ -62,6 +94,7 @@ export default {
   },
 
   mounted() {
+    this.updateChart();
     window.addEventListener("resize", this.resizeChart);
   },
